Add files via upload
Comparison of Region Based on Sales
Objective 
Help the organization by creating a dashboard to visualize the sales comparison between two selected regions. 
Description 
The director of a leading organization wants to compare the sales between two regions. He has asked each region’s operators to record the sales data to compare by region. The upper management wants to visualize the sales data using a dashboard to understand the performance between them and suggest the necessary improvements. 
Problem Statement 
MY task is to create a dashboard to understand the performance of the sales between two regions.
Task 
1. Create a hierarchy called Location for the variable Country. 
2. Create two parameters: Primary Region and Secondary Region 
3. Create a First Order Date 
4. Create a dashboard 
Partition the dashboard to display the details of the primary region and secondary region. 
• First Order Date 
• Total Sales 
• Average Sales per Order 
• No. of Customers 
• No. of Orders 
• No. of Products in Sale